+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 1.6.0
 
 - (Feature) When pressing `n` and `N` you can now navigate between search terms from the last search using `/`.
+- (Feature) Search for multiple terms using `/` separating them with `|` (the terms will be highlighted differently).
 
 ### 1.5.0